Open API - API와 DB related 설정(comment,BookmarkContent)
