+++ django_app/openapi/models/content.py
@@ -1,8 +1,10 @@
 from django.db import models
 
+from config import settings
+
 
 class Content(models.Model):
-    seq = models.CharField(max_length=2, unique=True)
+    seq = models.CharField(max_length=20, unique=True)
     title = models.CharField(max_length=100, null=True)
     start_date = models.DateField()
     end_date = models.DateField()
@@ -10,7 +12,7 @@ class Content(models.Model):
     realm_name = models.CharField(max_length=10, null=True)
     area = models.CharField(max_length=10, null=True)
     price = models.CharField(max_length=30, null=True)
-    content = models.CharField(max_length=5000, null=True)
+    content = models.TextField(max_length=5000, null=True)
     ticket_url = models.CharField(max_length=100, null=True)
     phone = models.CharField(max_length=50, null=True)
     thumbnail = models.CharField(max_length=200, null=True)
@@ -19,4 +21,4 @@ class Content(models.Model):
     place_url = models.CharField(max_length=200, null=True)
     place_addr = models.CharField(max_length=100, null=True)
     place_seq = models.CharField(max_length=10, null=True)
-    commant = models.CharField(max_length=500, null=True)
+    comment = models.ForeignKey(settings.AUTH_USER_MODEL, null=True)
