+++ django_app/openapi/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import Content
+from .models.content import Content
 
 
 class ContentSerializer(serializers.ModelSerializer):
@@ -8,4 +8,4 @@ class ContentSerializer(serializers.ModelSerializer):
         model = Content
         fields = ('seq', 'title', 'start_date', 'end_date', 'place', 'realm_name',
                   'area', 'price', 'content', 'ticket_url', 'phone', 'thumbnail',
-                  'place_url', 'place_addr')
+                  'place_url', 'place_addr', 'gps_x', 'gps_y')
